Move gconf's --enable-static from modulesets to jhbuildrc
GConf needs --enable-static because it has a test which tries to ld a .so bundle. That works fine on Linux, but not on OSX, so linking it statically allows the test to run and the build to complete without intervention.

The global autogenargs applies its arguments after the moduleset's args (makes sense, otherwise it wouldn't be able to override them), so putting --enable-static in the module didn't work.