Move python from bootstrap to gtk-osx-python-modules
There's generally no need to build python unless you're building gtk-osx-python, so having it in bootstrap is superfluous. Even if you're building gtk-osx-python, the python that came with OSX may be adequate (it's not if you're building on Tiger), but if you're going to distribute a bundle we recommend including python in the bundle to avoid version conflicts.