Create separate build chain for gtk3
This rather involved changeset separates out the dependencies needed for gtk2 and gtk3 (because of version differences) and creates new "meta-gtk-osx-gtk3" and "meta-gtk-osx-python-gtk3" meta-modules to build the correct set of dependencies.

The gtk+-3.0 module, added earlier, now includes the patches that haven't been approved or weren't included in the 3.2 release at the beginning of October.

While the two should be mutually installable (gtk2 and gtk3 and pygobject2 and pygobject3 have different-named installation directories and pkgconfig module names), several of the dependencies don't recognize that different versions are needed for each, so it's advisable to install into separate prefixes.