Get Devhelp to build in modulesets-stable
Here are three patches for getting Devhelp to build in
modulesets-stable. First, Devhelp depends on libunique, which I
have added to gtk-osx-random.modules. Libunique needs a patch to
run properly, see bgo#663913, since otherwise it crashes when
Devhelp starts.

Then, the patch from bgo#646962 corrects a syntax error in Devhelp,
and the patch from bgo#663789 ports Devhelp to GtkOSXApplication.

There is also an issue with the libtool script in the Devhelp
tarball. For more info, see the pull request. For that reason
mainly, I have changed the moduleset entry to
autogen-sh="autogen.sh" instead of "configure".