Rearrange bootstrap.modules and gtk-osx-bootstrap.modules
So that all tools are in the former and all dependencies are in the latter. The intention is to enable building the tools into ~/.local once instead of requiring an individual build per tree.

That doesn't work quite right because xz at least won't build universal 32/64 directly (I haven't yet tried separate compiles and lipo).