Javascript metamodule
This adds a new modules file with a target meta-gtk-osx-javascript
which builds GJS and its dependencies (mozjs17, nspr).

This target exists in modulesets-stable (GJS 1.38.1) and modulesets
(gnome-3-10 branch), but not yet in modulesets-unstable; I haven't
gotten mozjs24 to build yet. In both cases, mozjs17 and nspr are built
from tarballs.