Update GTK to 3.14, adjust patches and deps
This updates GTK to 3.14 on modulesets-stable and modulesets. On stable,
this obsoletes one patch, but needs another one; only for this minor
release, though, as it's already fixed on master and backported to the
stable branch.

On modulesets-unstable, the 3.15 version of GTK now has the start of
OpenGL support on the X11 and Wayland backends. That does not yet exist
on the Quartz backend, but even so libepoxy is now a build dependency.

Libepoxy, unfortunately, has a hardcoded dependency on X11. This libepoxy
module points at a pull request in Github which removes that dependency.
We still do need a package called xorg-util-macros, but that isn't a
heavy dependency and doesn't introduce any other dependencies itself.