Build DevHelp
DevHelp is a really nice tool to have available, so let's get it to
build! The latest version depends on WebKit2 (the webkit2gtk3 module
which doesn't work yet) so first of all we have to pin it to the last
version that supported WebKit1 (module name webkit1gtk3).

In addition, there are some M4 macros in that version that aren't
properly set up, so building from Git is broken. That means that we use
the tarball on modulesets as well.

It requires gsettings-desktop-schemas, for which we add a module.

Also, it will build without shared-mime-info, but then the doc pages
will display as HTML source code, rather than HTML. So effectively,
shared-mime-info is a dependency as well.