Build WebKitGTK 2.10.2
This adds a "webkit2gtk3" module. (Rationale for the name, discussed at
the time that "webkit1gtk3" was added: this is WebKitGTK's WebKit2 API
for GTK 3.)

WebKit has replaced its autotools build system with a cmake one, and so
this is a cmake module. They dropped support for the Quartz backend, so
the first patch re-enables that. The second patch fixes various build
issues stemming from the assumption that building on an OS X system means
you are compiling Webkit for Safari. The other three patches fix minor
build issues that I am still trying to get the WebKit maintainers to
merge.