Don't link XQuartz version of freetype2
As the freetype2 library isn't a dependency of cairo now, it's not
built by jhbuild. As a consequence, when cairo build script doesn't
find it, it looks further in the system and when XQuartz is installed,
it finds /opt/X11/bin/freetype-config and uses the XQuartz version
of the library.

This patch removes /opt/X11/bin from PATH to make sure the
XQuartz version isn't found and that the resulting binaries don't
depend on external libraries.