Optionally build bash as part of gtk-osx-bootstrap, skipped by default.
OS X 10.11 implements something Apple calls System Integrity Protection,
or SIP. One of its somewhat misguided behaviors is to strip DYLD_* and
LD_* environment variables from forked/execed processes if the parent is
a system process, and system process is defined as among other things
one whose executable path is in /bin or /usr/bin. This breaks things
that e.g. need to search for built-but-not-installed libraries or
libraries whose installed names aren't a full path.

The workaround is to build or copy the problem executables into some
other path and make sure that they're what's used. This change will
build bash in the $PREFIX directory; any dependent shell scripts in the
build system will need to be edited to use that bash in their shebang
instead of /bin/sh.