Better work-around for El Capitan's SIP:
Move bash to bootstrap.modules so that it's built as part of the
bootstarp step. For all other modulesets export CONFIG_SHELL pointing to
that bash, which ensures that everything with @SHELL@ in its .in files
along with libtool .la files and automake shell scripts like test-driver
get that path instead of /bin/sh, which eats the linker environment on
El Cap.