Avoid passing extra autogenargs to openssl
Openssl will error out if passed arguments to its configure script that
it doesn't understand. This is a problem when setting disable_Werror in
jhbuildrc, because that will cause --disable-Werror to be passed to all
autotools modules, after any chance we may have to edit the module's
autogenargs. Luckily we can negate any extra args that jhbuild may try to
pass by adding a # character.