Use SDKDIR=/ when building with command-line tools.
The SDK installed with command-line tools is th ecurrent
latest-and-greatest regardless of the installed MacOS version. Autotools
feature detection will notice that a symbol is available without
recognizing that it isn't available for macosx-version-min, causing link
failures later. The libraries installed in /usr/lib are adjusted to
report only the symbols available for the running MacOS version.