Add a gtk.14 module to modulesets-stable/gtk-osx.modules
We've discovered on Gramps that versions of Gtk+ newer than gtk+-3.14
are quite unstable, at least in a pygobject environment. This module
makes it easier to build a stable bundle for similarly impacted applications.

A companion meta-module, meta-gtk-osx-gtk3.14 is included, simply adjust
your modules list. Do note that if you have relied in the past on certain
upstream modules to "pull in" meta-gtk-osx-gtk3 that their dependencies
have been made "soft" in order to build with either so you'll need to
explicitly add meta-gtk-osx-gtk3 to ensure that everything is completely
built.