Update installation of jhbuildrc and jhbuildrc-custom...
To match the current jhbuild defaults.
Notice that one can now set XDG_CONFIG_HOME and put both there, and it
need not be in $HOME. The default is $HOME/.config. $HOME/.jhbuildrc and
$HOME/.jhbuildrc-custom are still supported and won't be disturbed if
found.

jhbuildrc will now search in order the same directory it is in,
$XDG_CONFIG_HOME if it's set or $HOME/.config if its not, and $HOME for
either jhbuildrc-custom or .jhbuildrc-custom.
If for some reason you want jhbuildrc-custom somewhere else or to have a
different name you can set $JHBUILDRC_CUSTOM. Both gtk-osx-setup.sh and
jhbuildrc will honor it.