Only pip-uninstall packages that we installed.
If a package (in particular typing and enum34) is found in the system
python directories pip will notice and not install it. If we later
tell pip to uninstall it it will try and of course fail to remove it
from the system directory. Check first that the package is installed in
our directory before trying to uninstall it.