Reconfigure harfbuzz and freetype as dependencies.
Harfbuzz and Freetype have become mandatory dependencies via Pango and
librsvg. Freetype in turn has a circular dependency on Harbuzz, but
can be built without it. Quartz doesn't really need harfbuzz for shaping
we just need to satisfy the dependencies. So harfbuzz is moved out of
meta-gtk-osx-freetype into gtk-osx.modules as a dependency of pango. It
needs cairoft2 so cairo gets freetype-no-harfbuzz as an added depenedncy
as well, as does librsvg.

In the process upgraded both freetype modules and converted them to build
with CMake, upgraded harfbuzz and fontconfig and converted them to build
with meson on all three modulesets,