Remove meta-gtk-osx-freetype, make modules part of core metamodules.
The basic problem is two circular dependencies in the text handling
modules:
1. freetype2 depends on harfbuzz and harfbuzz depends on freetype2.
2. harbuzz depends on gobject-introspection which depends on cairo which
depends on harfbuzz.

One more little detail: If pango detects freetype it insists on
fontconfig as well, and fontconfig needs the freetype built with
harfbuzz. Both of those need to be built before cairo because pango
wants the cairoft2 and cairofc. In that case pango also demands harfbuzz
with freetype2 support and with gobject introspection.

librsvg, required for icon rendering, requires freetype2 but doesn't
need any of the others.

It proved unreasonably complex to have meta-gtk-osx-freetype as an
optional module so remove it (a stub remains to prevent breaking
modulesets that use it).