Don't try to build gnutls with Xcode older than 10.0
gnutls built with p11-kit support has a weird side effect of breaking
g-ir-scanner when using an older Xcode. glib-networking requires
gnutls to be built with p11-kit, so if Xcode is older than 10.0
don't build gnutls at all. We always build openssl and glib-networking
with openssl support which should be sufficient.