Improve setting Rust environment variables.
If rustup is found, see if it's in DEVPREFIX; if so figure out where
so that defaults for CARGO_HOME and RUSTUP_HOME will have a better
chance of being correct. If it's not default them to the normal rust
installation defaults. If they're different they should already be set up.