Add pygments as a dependency for setting python version.
Gtk-doc requires the python3 that builds it to have pygments available,
so we have a pygments module that depends on python3. It's a soft
dependency for gtk-doc so that one can build it with the virtenv python,
but if one needs python3, say to bundle the app, then specify pygments
to pull in python3. This change will recognize it and set the build
environment python accordingly.