Fix building libxml2
This is in two parts. The patch is for
https://gitlab.gnome.org/GNOME/libxml2/-/issues/317. The rest is
because while Apple provides python3 as part of Xcode 12 and 13 for
macOS 11 and 12 respectively it doesn't provide a linkable libpython3.dylib so
we can't build the python bindings for libxml2 or gobject-introspection
with it. To resolve that we force install a pyenv python3 and set up
jhbuild accordingly, and add an LDFLAGS entry to locate its library.