Replace pkg-config with pkgconf.
pkg-config is no longer being maintained and its ancient glib has caught
up with it. Many distros including RHEL and MSYS2 have already made the
replacement.

We need a symlink to pkg-config for compatibility with some packages that
attempt to call it directly instead of using autotools/cmake/meson macros.

pkgconf is a bit absolute about not emitting -L in --libs output if it
considers the path a system path, and its default system path is (prefix).
We tell it to use /usr/lib and /usr/include as the system paths instead so
that it will emit -L(prefix)/lib and -I(prefix)/inxclude.