Add CMAKE_FIND_FRAMEWORK=LAST globally
This prevents issues with header files being found in frameworks like the Mono framework which are incompatible with the dylibs in the gtk/inst/lib folder. An example is libjpeg, which would read the JPEG_LIB_VERSION from the Mono framework and take the dylib with a different version from the gtk/inst/lib folder.