Annual mass update: gtk-osx.modules
libxml2 is now a dependency of librsvg so move it to gtk-osx.modules.
Switched to cmake instead of meson because meson just wraps cmake and
we need to set CMAKE_MACOS_RPATH=NO because with it on it sets the
install-name to @rpath but doesn't set an RPATH in xmllint.